+++ src/main/java/codesquad/fineants/spring/api/portfolio_stock/PortfolioStockService.java
@@ -22,6 +22,11 @@ import codesquad.fineants.spring.api.errors.exception.ForBiddenException;
 import codesquad.fineants.spring.api.errors.exception.NotFoundResourceException;
 import codesquad.fineants.spring.api.kis.manager.CurrentPriceManager;
 import codesquad.fineants.spring.api.kis.manager.LastDayClosingPriceManager;
+import codesquad.fineants.spring.api.portfolio_stock.chart.DividendChart;
+import codesquad.fineants.spring.api.portfolio_stock.chart.PieChart;
+import codesquad.fineants.spring.api.portfolio_stock.chart.SectorChart;
+import codesquad.fineants.spring.api.portfolio_stock.factory.PortfolioDetailFactory;
+import codesquad.fineants.spring.api.portfolio_stock.factory.PortfolioHoldingDetailFactory;
 import codesquad.fineants.spring.api.portfolio_stock.request.PortfolioStockCreateRequest;
 import codesquad.fineants.spring.api.portfolio_stock.response.PortfolioChartResponse;
 import codesquad.fineants.spring.api.portfolio_stock.response.PortfolioDetailRealTimeItem;
